requirements.txt
groq==0.4.2
python-dotenv>=1.0.0
pydantic>=2.7.0
requests>=2.31.0